+++ include/maec/audio_buffer.hpp
@@ -4,9 +4,9 @@
  * @brief Components for AudioBuffers
  * @version 0.1
  * @date 2022-11-16
- * 
+ *
  * @copyright Copyright (c) 2022
- * 
+ *
  * This file contains components for working
  * with Buffers that hold audio.
  * We provide some components for converting
@@ -21,7 +21,7 @@
 #include "dsp/buffer.hpp"
 
 /// A typedef representing an AudioBuffer
-using AudioBuffer = Buffer<long double>;
+using AudioBuffer = Buffer<double>;
 
 /// Alias for a unique pointer to an AudioBuffer
 using BufferPointer = std::unique_ptr<AudioBuffer>;
@@ -42,16 +42,17 @@ std::unique_ptr<AudioBuffer> create_buffer(int size, int channels);
 
 /**
  * @brief Create an AudioBuffer
- * 
+ *
  * Creates a buffer object and returns the result.
  * The resulting pointer will be unique, and has ownership of the buffer.
- * 
+ *
  * @param size Size of buffer to create
  * @param channels Channels in AudioBuffer
  * @param sample_rate Sample rate of audio data
  * @return std::unique_ptr<AudioBuffer> Newly created buffer
  */
-std::unique_ptr<AudioBuffer> create_buffer(int size, int channels, double sample_rate);
+std::unique_ptr<AudioBuffer> create_buffer(int size, int channels,
+                                           double sample_rate);
 
 /**
  * @brief Components that squish and split audio buffers.
@@ -78,36 +79,39 @@ std::unique_ptr<AudioBuffer> create_buffer(int size, int channels, double sample
  * Most squishers will take a vector and convert it into an AudioBuffer.
  *
  * Squishers and splitters support format conversion,
- * which means that the format of the incoming container (be it long double, float, int, ect)
- * can be converted into another format during the copy operation.
- * This saves us some time, as otherwise we would have do two copy operations.
- * We can just go ahead and get both out of the way in one operation.
+ * which means that the format of the incoming container (be it double,
+ * float, int, ect) can be converted into another format during the copy
+ * operation. This saves us some time, as otherwise we would have do two copy
+ * operations. We can just go ahead and get both out of the way in one
+ * operation.
  */
 
 /**
  * @brief Converts a given AudioBuffer into a squished interleaved vector
- * 
+ *
  * We take the data contained in the audio buffer and squish it.
  * The final vector will be in interleaved format.
- * 
+ *
  * We require an iterator that is the destination
  * for this audio data.
- * 
+ *
  * @param buff AudioBuffer to work with
  * @param iter Input iterator of the output vector
  */
 template <typename It, typename Func>
-void squish_inter(AudioBuffer* buff, const It& iter, Func oper) { std::transform(buff->ibegin(), buff->iend(), iter, oper); }
+void squish_inter(AudioBuffer* buff, const It& iter, Func oper) {
+    std::transform(buff->ibegin(), buff->iend(), iter, oper);
+}
 
 /**
  * @brief Converts a given AudioBuffer into a squished sequential vector
- * 
+ *
  * We take the data contained in the audio buffer and squish it.
  * The final vector will be in sequential format.
- * 
+ *
  * We require the iterator that is the destination
  * for this audio data.
- * 
+ *
  * @param buff AudioBuffer to work with
  * @param iter Input iterator of the output vector
  */
@@ -127,36 +131,38 @@ void squish_seq(AudioBuffer* buff, const It& iter, Func oper) {
  * @param iter Input iterator to ignore
  */
 template <typename It, typename Func>
-void squish_null(AudioBuffer* buff, const It& iter, Func oper) {}  // NOLINT(clang-diagnostic-unused-parameter)
+void squish_null(AudioBuffer* buff, const It& iter, Func oper) {
+}  // NOLINT(clang-diagnostic-unused-parameter)
 
 /**
  * @brief Components for converting audio data
- * 
+ *
  * This section contains components for converting audio data.
  * These are used to convert audio data from one format to another.
- * Most audio libraries do not work well with the internal maec audio format, long double,
- * so it is necessary to provide components for transforming between the two.
- * 
+ * Most audio libraries do not work well with the internal maec audio format,
+ * double, so it is necessary to provide components for transforming
+ * between the two.
+ *
  * It is recommended to use these functions with the squishers/splitters
  * defined above, which will automatically iterate over the provided array
  * and convert it into the necessary data structure.
- * 
+ *
  * The naming convention goes like this:
- * 
+ *
  * For converting the maec format (mf) into another:
- * 
- * [OTHER] mf_[OTHER](long double val)
- * 
+ *
+ * [OTHER] mf_[OTHER](double val)
+ *
  * For converting other format into mf:
- * 
- * long double [OTHER]_mf([OTHER] val)
- * 
+ *
+ * double [OTHER]_mf([OTHER] val)
+ *
  * Where [OTHER] is the name of the other format to work with.
  * For example, converting mf to int and vice versa:
- * 
- * mf_int(long double val)
+ *
+ * mf_int(double val)
  * int_mf(int val)
- * 
+ *
  * It is also worth mentioning that 'null' means the function will do nothing.
  * For example, mf_null() will just return what is given.
  */
@@ -165,24 +171,24 @@ void squish_null(AudioBuffer* buff, const It& iter, Func oper) {}  // NOLINT(cla
 
 /**
  * @brief Converts mf to a float
- * 
+ *
  * This function is simply a cast!
  * There are no interesting methods utilized here.
- * 
+ *
  * @param val Value to convert
  * @return float Converted float value
  */
-float mf_float(long double val);
+float mf_float(double val);
 
 /**
  * @brief Does nothing
- * 
+ *
  * This function simply returns the value provided.
- * 
+ *
  * @param val Value to convert
- * @return long double Value provided
+ * @return double Value provided
  */
-long double mf_null(long double val);
+double mf_null(double val);
 
 /**
  * @brief Converts mf to a 16bit integer
@@ -193,43 +199,43 @@ long double mf_null(long double val);
  * @param val Value to convert
  * @return int16_t Converted int16_t
  */
-int16_t mf_int16(long double val);
+int16_t mf_int16(double val);
 
 /**
  * @brief Converts mf to a unsigned 16bit integer
- * 
+ *
  * We multiply the mf by 32767,
  * flip the most significant bit,
  * and cast it to uint16_t
- * 
- * @param val Value to convert 
+ *
+ * @param val Value to convert
  * @return uint16_t Converted uint16_t
  */
-uint16_t mf_uint16(long double val);
+uint16_t mf_uint16(double val);
 
 /**
  * @brief Converts mf to a char
- * 
+ *
  * We multiply the mf by 127
  * and cast it to char.
- * 
+ *
  * @param val Value to convert
  * @return char Converted uint16_t
  */
-char mf_char(long double val);
+char mf_char(double val);
 
 /**
  * @brief Converts mf to an unsigned char
- * 
+ *
  * We add one to the given value
  * Divide the result by 2,
  * multiply by 255,
  * and then round.
- * 
+ *
  * @param val Value to convert
  * @return uchar Converted uchar
  */
-unsigned char mf_uchar(long double val);
+unsigned char mf_uchar(double val);
 
 // other -> mf
 
@@ -240,51 +246,51 @@ unsigned char mf_uchar(long double val);
  * and then cast to mf
  *
  * @param val Value to convert
- * @return long double Converted mf
+ * @return double Converted mf
  */
-long double int16_mf(int16_t val);
+double int16_mf(int16_t val);
 
 /**
- * @brief Converts an unsigned 16bit integer to mf 
+ * @brief Converts an unsigned 16bit integer to mf
  *
  * We first normalize the value by 65535,
  * double it, and then subtract by 1.
- * 
+ *
  * @param val Value to convert
- * @return long double 
+ * @return double
  */
-long double uint16_mf(uint16_t val);
+double uint16_mf(uint16_t val);
 
 /**
  * @brief Converts a char to mf
- * 
+ *
  * We simply cast the char to mf
  * and then divide by 127;
- * 
+ *
  * @param val Value to convert
- * @return long double Converted mf
+ * @return double Converted mf
  */
-long double char_mf(char val);
+double char_mf(char val);
 
 /**
  * @brief Converts a unsigned char to mf
- * 
+ *
  * We first normalize the value by 255,
- * double it, and then subtract 1. 
- * 
+ * double it, and then subtract 1.
+ *
  * @param val Value to convert
- * @return long double Converted mf
+ * @return double Converted mf
  */
-long double uchar_mf(unsigned char val);
+double uchar_mf(unsigned char val);
 
 /**
  * @brief Components for byte conversions
- * 
+ *
  * This section describes functions for converting
  * byte data into higher level types.
  * These functions operate with character arrays,
  * and pointers/iterators to specific elements should be provided!
- * 
+ *
  * These functions follow the same naming convention mentioned above.
  */
 
@@ -292,18 +298,18 @@ long double uchar_mf(unsigned char val);
 
 /**
  * @brief Converts byte data into 16bit integers
- * 
+ *
  * This function utilizes endian safe methods
  * for conversions.
- * 
+ *
  * We require 2 bytes to make this conversion,
  * so your pointer should have at least 1 value in front of it.
- * 
+ *
  * @tparam T Iterator to character data
  * @param byts Byte data to convert
  * @return int16_t 16bit integer
  */
-template<typename T>
+template <typename T>
 int16_t char_int16(T byts) {
 
     // Define the type:
@@ -322,17 +328,17 @@ int16_t char_int16(T byts) {
 
 /**
  * @brief Converts bytes data into 32bit integers
- * 
+ *
  * This function utilizes endian safe methods for conversions.
- * 
+ *
  * We require 4 bytes to make this conversion,
  * so your pointer should have at least 3 values in front of it.
- * 
+ *
  * @tparam T Iterator to byte data
  * @param byts Bytes data to convert
  * @return int32_t 32bit integer
  */
-template<typename T>
+template <typename T>
 int32_t char_int32(T byts) {
 
     // Define the value:
@@ -356,18 +362,18 @@ int32_t char_int32(T byts) {
 
 /**
  * @brief Converts byte data into unsigned 32bit integers
- * 
+ *
  * This function utilizes endian safe methods for conversions,
  * and in fact is identical to char_int32, except this is the unsigned variant.
- * 
+ *
  * We require 4 bytes to make this conversion.
  * so your pointer should have at least 3 values in front of it.
- * 
+ *
  * @tparam T Iterator to byte data
  * @param byts Byte data to convert
  * @return uint32_t unsigned 32bit integer
  */
-template<typename T>
+template <typename T>
 uint32_t char_uint32(T byts) {
 
     // Define the value:
@@ -393,18 +399,18 @@ uint32_t char_uint32(T byts) {
 
 /**
  * @brief Converts a 16bit integer into byte data
- * 
+ *
  * This function utilizes endian safe methods for conversions.
- * 
+ *
  * We will place the result into the provided iterable.
  * We require 2 bytes to make this conversion, so your pointer should
  * have space for 2 values!
- * 
+ *
  * @tparam T Iterator type of output byte data
  * @param val Value to convert
  * @param byts Iterator to output byte data
  */
-template<typename T>
+template <typename T>
 void int16_char(int16_t val, T byts) {
 
     // Write each part of the value:
@@ -415,18 +421,18 @@ void int16_char(int16_t val, T byts) {
 
 /**
  * @brief Converts a 32bit integer into byte data
- * 
+ *
  * This function utilizes endian safe methods for conversions.
- * 
+ *
  * We will place the result into the provided iterable.
  * We require 4 bytes to make this conversion, so your pointer should
  * have space for 4 values!
- * 
+ *
  * @tparam T Iterator type of output byte data
  * @param val Value to convert
  * @param byts Iterator to output byte data
  */
-template<typename T>
+template <typename T>
 void int32_char(int32_t val, T byts) {
 
     // Iterate a number of times
@@ -442,19 +448,19 @@ void int32_char(int32_t val, T byts) {
 
 /**
  * @brief Converts an unsigned 32bit integer into byte data
- * 
+ *
  * This function utilizes endian safe methods for conversions,
  * and in fact is identical to uint32_char, except this is the unsigned variant.
- * 
+ *
  * We will place the result into the provided iterable.
  * We require 4 bytes to make this conversion, so your pointer should
  * have space for 4 values!
- * 
+ *
  * @tparam T Iterator type of output byte data
  * @param val Value to convert
  * @param byts Iterator to output byte data
  */
-template<typename T>
+template <typename T>
 void uint32_char(uint32_t val, T byts) {
 
     // Just cast and call:
