+++ include/maec/dsp/kernel.hpp
@@ -4,19 +4,19 @@
  * @brief Tools for generating and manipulating kernels
  * @version 0.1
  * @date 2023-07-26
- * 
+ *
  * @copyright Copyright (c) 2023
- * 
+ *
  * A 'kernel' is a finite impulse response that is being used
  * to filter a signal via convolution.
  * Filtering a signal is to preform signal separation
  * or signal restoration.
- * 
+ *
  * Signal separation is the process of separating (or removing)
  * certain components to achieve an outcome.
  * Signal restoration is the process of restoring components
  * that may have been distorted or corrupted.
- * 
+ *
  * This file contains tools and components for generating
  * and manipulating these filter kernels,
  * allowing you to use them in your operations.
@@ -26,22 +26,22 @@
 
 #include <algorithm>
 
-#include "dsp/window.hpp"
 #include "dsp/util.hpp"
+#include "dsp/window.hpp"
 
-using window_functiont = long double(*)(int, int);
+using window_functiont = double (*)(int, int);
 
 /**
  * @brief Preforms spectral inversion on a filter kernel.
- * 
+ *
  * Spectral inversion flips the frequency response of
  * the given kernel top-for-bottom.
  * This has the process of inverting a filter to preform
  * the opposite operation:
- * 
+ *
  * High Pass <-> Low Pass
  * Band Reject <-> Band Pass
- * 
+ *
  * From a purely mathematical perspective, we are applying
  * the filter to two systems in parallel, negating the output of one,
  * and summing their results.
@@ -49,28 +49,29 @@ using window_functiont = long double(*)(int, int);
  * while the other does nothing.
  * Finally, the output is defined as the output of the all pass system
  * minus the output of the original filter system.
- * This means that the original frequency components are subtracted from the signal,
- * thus resulting in output that is the negation of the original filter kernel.
+ * This means that the original frequency components are subtracted from the
+ * signal, thus resulting in output that is the negation of the original filter
+ * kernel.
  *
  * This operation can be simplified by combining the two signal filters in
  * our theoretical system.
  * This leaves us with two steps:
- * 
+ *
  * 1. Negate the incoming filter
  * 2. Add 1 to the sample at the center of symmetry.
- * 
+ *
  * This process does have some caveats.
  * The incoming filter MUST have left-right symmetry,
  * and the filter kernel MUST be of an odd length.
- * 
+ *
  * This function works in place!
  * Simply provide your filter kernel and it will be inverted.
- * 
+ *
  * @tparam T Filter iterator type
  * @param input Iterator to filter data
  * @param size Size of filter kernel
  */
-template<typename I>
+template <typename I>
 void spectral_inversion(I input, int size) {
 
     // First, negate each sample:
@@ -79,12 +80,12 @@ void spectral_inversion(I input, int size) {
 
         // Negate this sample:
 
-        *(input+i) *= -1;
+        *(input + i) *= -1;
     }
 
     // Add one to the center sample:
 
-    *(input + ((size/ 2) + 1)) += 1;
+    *(input + ((size / 2) + 1)) += 1;
 }
 
 template <typename I>
@@ -101,14 +102,15 @@ void spectral_reversal(I input, int size) {
 }
 
 template <typename O>
-void sinc_kernel(double freq, int size, O output, window_functiont window = window_blackman) {
+void sinc_kernel(double freq, int size, O output,
+                 window_functiont window = window_blackman) {
 
     // Determine some constants:
 
     const int size2 = (size - 1) / 2;
-    const long double inner = 2 * M_PI * freq;
+    const double inner = 2 * M_PI * freq;
 
-    long double sum = 1.;
+    double sum = 1.;
 
     // Iterate over first half:
 
@@ -120,7 +122,7 @@ void sinc_kernel(double freq, int size, O output, window_functiont window = wind
 
         // Calculate value:
 
-        const long double value = sinc(inner * half) * window(i, size);
+        const double value = sinc(inner * half) * window(i, size);
 
         // Set value on each half:
